multicore support

+++ config/redis.js
@@ -4,7 +4,7 @@ var redisUrl = url.parse('redis://user:@localhost:6379/');
 
 var db = redis.createClient();
 
-if (process.env.FLUSH_REDIS === 'true')
+if (process.env.FLUSH_REDIS === 'true' && parseInt(process.env.NODE_APP_INSTANCE) === 0)
   db.flushdb();
 
 module.exports.redis = db;
